ep10

+++ graphql/resolvers/booking.js
@@ -3,7 +3,11 @@ const Event = require('../../models/event');
 const { transformBooking, transformEvent } = require('./merge');
 
 module.exports = {
-  bookings: async () => {
+  bookings: async (args, req) => {
+    if (!req.isAuth) {
+      throw new Error('Unauthenticated');
+    }
+
     try {
       const bookings = await Booking.find();
 
@@ -16,10 +20,14 @@ module.exports = {
       throw err;
     }
   },
-  bookEvent: async (args) => {
+  bookEvent: async (args, req) => {
+    if (!req.isAuth) {
+      throw new Error('Unauthenticated');
+    }
+
     const fetchedEvent = await Event.findOne({ _id: args.eventId });
     const booking = new Booking({
-      user: '5ff402f05e96c9e87afadee2',
+      user: req.userId,
       event: fetchedEvent
     });
     
@@ -27,12 +35,14 @@ module.exports = {
 
     return transformBooking(result);
   },
-  cancelBooking: async (args) => {
+  cancelBooking: async (args, req) => {
+    if (!req.isAuth) {
+      throw new Error('Unauthenticated');
+    }
+
     try {
       const booking = await Booking.findById(args.bookingId).populate('event');
-
       const event = transformEvent(booking.event);
-
       await Booking.deleteOne({ _id: args.bookingId });
 
       return event;
@@ -42,6 +52,5 @@ module.exports = {
       throw err;
     }
   }
-
 
 }
